Removed ParamTwo F string since this is the carry flag.  Carry flag should be used as part of specific op codes and modified as such so there should be no need to map it.  We'll know exactly how and when to use/modify it.  Should add modifications as per insturction action.
Corrected LD instruction StringMAP that was including the HALT action

AF is not used as a 2nd parameter, removed

Talk about pushing Gameboy.Start() contents for running the OPCode instructions into a separate method because we need to execute an insturction before we can disable/enable interrupts.  Also depending on how we do the registers/memory that means we'll need to be passed in the memory locations so we can actually access them correctly.

DI,EI,HALT,STOP don't have any actual implementation

Think we'll need to change the GetInstruction and have it pass in a processor to the specific

SBC - Need to discuss Carry Flags
CP - Not fully implemented due to flags